+++ src/Hangfire.Mongo.Tests/MongoDistributedLockFacts.cs
@@ -15,14 +15,12 @@ namespace Hangfire.Mongo.Tests
     [Collection("Database")]
     public class MongoDistributedLockFacts
     {
-        private readonly IDistributedLockMutex _distributedLockMutex = new DistributedLockMutex();
         private readonly HangfireDbContext _database = ConnectionUtils.CreateDbContext();
         [Fact]
         public void Ctor_ThrowsAnException_WhenResourceIsNull()
         {
             var exception = Assert.Throws<ArgumentNullException>(
-                () => new MongoDistributedLock(null, TimeSpan.Zero, _database, new MongoStorageOptions(),
-                    _distributedLockMutex));
+                () => new MongoDistributedLock(null, TimeSpan.Zero, _database, new MongoStorageOptions()));
 
             Assert.Equal("resource", exception.ParamName);
         }
@@ -32,7 +30,7 @@ namespace Hangfire.Mongo.Tests
         {
             var exception = Assert.Throws<ArgumentNullException>(
                 () => new MongoDistributedLock("resource1", TimeSpan.Zero, null,
-                    new MongoStorageOptions(), _distributedLockMutex));
+                    new MongoStorageOptions()));
 
             Assert.Equal("dbContext", exception.ParamName);
         }
@@ -40,8 +38,7 @@ namespace Hangfire.Mongo.Tests
         [Fact, CleanDatabase]
         public void Ctor_SetLock_WhenResourceIsNotLocked()
         {
-            var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions(),
-                _distributedLockMutex);
+            var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions());
             using (lock1.AcquireLock())
             {
                 var locksCount =
@@ -54,8 +51,7 @@ namespace Hangfire.Mongo.Tests
         [Fact, CleanDatabase]
         public void Ctor_SetReleaseLock_WhenResourceIsNotLocked()
         {
-            var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions(),
-                _distributedLockMutex);
+            var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions());
             using (lock1.AcquireLock())
             {
                 var locksCount =
@@ -73,8 +69,7 @@ namespace Hangfire.Mongo.Tests
         [Fact, CleanDatabase]
         public void Ctor_AcquireLockWithinSameThread_WhenResourceIsLocked()
         {
-            var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions(),
-                _distributedLockMutex);
+            var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions());
             using (lock1.AcquireLock())
             {
                 
@@ -83,8 +78,7 @@ namespace Hangfire.Mongo.Tests
                         Builders<DistributedLockDto>.Filter.Eq(_ => _.Resource, "resource1"));
                 Assert.Equal(1, locksCount);
 
-                var lock2 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions(),
-                    _distributedLockMutex);
+                var lock2 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions());
                 using (lock2.AcquireLock())
                 {
                     locksCount =
@@ -98,8 +92,7 @@ namespace Hangfire.Mongo.Tests
         [Fact, CleanDatabase]
         public void Ctor_ThrowsAnException_WhenResourceIsLocked()
         {
-            var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions(),
-                _distributedLockMutex);
+            var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions());
             using (lock1.AcquireLock())
             {
                 var locksCount =
@@ -112,8 +105,7 @@ namespace Hangfire.Mongo.Tests
                     Assert.Throws<DistributedLockTimeoutException>(() =>
                         {
                             var lock2 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database,
-                                new MongoStorageOptions(),
-                                _distributedLockMutex);
+                                new MongoStorageOptions());
                             lock2.AcquireLock();
                         }
                         );
@@ -128,8 +120,7 @@ namespace Hangfire.Mongo.Tests
         {
             var t = new Thread(() =>
             {
-                var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions(),
-                    _distributedLockMutex);
+                var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, new MongoStorageOptions());
                 using (lock1.AcquireLock())
                 {
                     Thread.Sleep(TimeSpan.FromSeconds(3));
@@ -143,7 +134,7 @@ namespace Hangfire.Mongo.Tests
             // Record when we try to aquire the lock
             var startTime = DateTime.UtcNow;
             var lock2 = new MongoDistributedLock("resource1", TimeSpan.FromSeconds(10), _database,
-                new MongoStorageOptions(), _distributedLockMutex);
+                new MongoStorageOptions());
             using (lock2.AcquireLock())
             {
                 Assert.InRange(DateTime.UtcNow - startTime, TimeSpan.Zero, TimeSpan.FromSeconds(5));
@@ -155,8 +146,7 @@ namespace Hangfire.Mongo.Tests
         {
             var exception = Assert.Throws<ArgumentNullException>(() =>
             {
-                var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, null,
-                    _distributedLockMutex);
+                var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database, null);
                 lock1.AcquireLock();
             });
 
@@ -167,8 +157,7 @@ namespace Hangfire.Mongo.Tests
         public void Ctor_SetLockExpireAtWorks_WhenResourceIsNotLocked()
         {
             var lock1 = new MongoDistributedLock("resource1", TimeSpan.Zero, _database,
-                new MongoStorageOptions() {DistributedLockLifetime = TimeSpan.FromSeconds(3)},
-                _distributedLockMutex);
+                new MongoStorageOptions() {DistributedLockLifetime = TimeSpan.FromSeconds(3)});
             using (lock1.AcquireLock())
             {
                 DateTime initialExpireAt = DateTime.UtcNow;
@@ -189,8 +178,7 @@ namespace Hangfire.Mongo.Tests
             _database.DistributedLock.InsertOne(new DistributedLockDto {ExpireAt = initialExpireAt, Resource = "resource1" });
 
             var lock1 = new MongoDistributedLock("resource1", TimeSpan.FromSeconds(5), _database,
-                new MongoStorageOptions(),
-                _distributedLockMutex);
+                new MongoStorageOptions());
             using (lock1.AcquireLock())
             {
                 var lockEntry = _database.DistributedLock