+++ src/Hangfire.Mongo.Tests/MongoNotificationObserverFacts.cs
@@ -15,15 +15,15 @@ namespace Hangfire.Mongo.Tests
         private readonly HangfireDbContext _dbContext;
 
         private readonly Mock<IJobQueueSemaphore> _jobQueueSemaphoreMock;
-        private readonly Mock<IDistributedLockMutex> _distributedLockMutexMock;
         private readonly CancellationTokenSource _cts;
         public MongoNotificationObserverFacts()
         {
             _dbContext = ConnectionUtils.CreateDbContext();
             _jobQueueSemaphoreMock = new Mock<IJobQueueSemaphore>(MockBehavior.Strict);
-            _distributedLockMutexMock = new Mock<IDistributedLockMutex>(MockBehavior.Strict);
-            var mongoNotificationObserver = new MongoNotificationObserver(_dbContext, _jobQueueSemaphoreMock.Object,
-                _distributedLockMutexMock.Object);
+            var mongoNotificationObserver = new MongoNotificationObserver(
+                _dbContext,
+                new MongoStorageOptions(),
+                _jobQueueSemaphoreMock.Object);
             
             _dbContext.Database.DropCollection(_dbContext.Notifications.CollectionNamespace.CollectionName);
             var migration = new AddNotificationsCollection();
@@ -58,24 +58,6 @@ namespace Hangfire.Mongo.Tests
             
             // ASSERT
             _jobQueueSemaphoreMock.Verify(m => m.Release("test"), Times.Once);
-            _distributedLockMutexMock.Verify(m => m.Release(It.IsAny<string>()), Times.Never);
-        }
-        
-        [Fact]
-        public void Execute_LockReleasedEnqueued_Signaled()
-        {
-            // ARRANGE
-            var signal = new SemaphoreSlim(0,1);
-            _jobQueueSemaphoreMock.Setup(m => m.Release("test"))
-                .Callback(() => signal.Release());
-            
-            // ACT
-            _dbContext.Notifications.InsertOne(NotificationDto.LockReleased("test"));
-            signal.Wait(1000);
-            
-            // ASSERT
-            _jobQueueSemaphoreMock.Verify(m => m.Release(It.IsAny<string>()), Times.Never);
-            _distributedLockMutexMock.Verify(m => m.Release("test"), Times.Once);
         }
     }
 }